refactor(types): move SearchResponseRequestParams interface to Types.ts
- Move `SearchResponseRequestParams` interface from `Documents.ts` to `Types.ts`
- Update imports in `Documents.ts` to include the moved interface
- Re-export the interface from `Documents.ts` to maintain backward compatibility
