+++ src/Typesense/Documents.ts
@@ -3,7 +3,11 @@ import ApiCall from "./ApiCall";
 import Configuration from "./Configuration";
 import { ImportError, RequestMalformed } from "./Errors";
 import { SearchOnlyDocuments } from "./SearchOnlyDocuments";
-import { SearchParams, WriteableDocuments } from "./Types";
+import {
+  SearchParams,
+  SearchResponseRequestParams,
+  WriteableDocuments,
+} from "./Types";
 
 // Todo: use generic to extract filter_by values
 export type DeleteQuery =
@@ -90,17 +94,6 @@ export interface SearchResponseFacetCountSchema<T extends DocumentSchema> {
     max?: number;
     min?: number;
     sum?: number;
-  };
-}
-
-export interface SearchResponseRequestParams {
-  collection_name?: string;
-  q?: string;
-  page?: number;
-  per_page?: number;
-  first_q?: string;
-  voice_query?: {
-    transcribed_query?: string;
   };
 }
 
@@ -405,4 +398,5 @@ export type {
   UnionArraySearchParams,
   ArraybleParams,
   ExtractBaseTypes,
+  SearchResponseRequestParams,
 } from "./Types";
